refactor: убрать DirectorDbStorage из FilmDbStorage, style: удалить неиспользуемые импорты
