+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -1,7 +1,7 @@
 package ru.yandex.practicum.filmorate.service;
 
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.enums.SearchingParts;
 import ru.yandex.practicum.filmorate.exceptions.InvalidIdException;
@@ -9,9 +9,10 @@ import ru.yandex.practicum.filmorate.exceptions.filmExceptions.BadSearchQueryExc
 import ru.yandex.practicum.filmorate.exceptions.filmExceptions.LikesException;
 import ru.yandex.practicum.filmorate.model.Feed;
 import ru.yandex.practicum.filmorate.model.Film;
+import ru.yandex.practicum.filmorate.storage.director.DirectorStorage;
 import ru.yandex.practicum.filmorate.storage.feed.FeedStorage;
-import ru.yandex.practicum.filmorate.storage.film.FilmDbStorage;
-import ru.yandex.practicum.filmorate.storage.user.UserDbStorage;
+import ru.yandex.practicum.filmorate.storage.film.FilmStorage;
+import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 import ru.yandex.practicum.filmorate.validators.FilmValidator;
 
 import java.util.*;
@@ -19,22 +20,17 @@ import java.util.*;
 
 @Slf4j
 @Service
+@RequiredArgsConstructor
 public class FilmService {
 
-    private final FilmDbStorage filmStorage;
-    private final UserDbStorage userStorage;
+    private final FilmStorage filmStorage;
+    private final UserStorage userStorage;
     private final FeedStorage feedStorage;
+
+    private final DirectorStorage directorStorage;
     private final FilmValidator filmValidator;
 
-    @Autowired
-    public FilmService(FilmDbStorage filmStorage, UserDbStorage userStorage, FilmValidator filmValidator, FeedStorage feedStorage) {
-        this.filmStorage = filmStorage;
-        this.userStorage = userStorage;
-        this.filmValidator = filmValidator;
-        this.feedStorage = feedStorage;
-    }
-
-    public void addLikeToFilm(int filmId, int userId)  {
+    public void addLikeToFilm(int filmId, int userId) {
         userStorage.findUserById(userId).orElseThrow(
                 () -> new InvalidIdException("Пользователь с id" + userId + " не найден"));
         filmStorage.findFilmById(filmId).orElseThrow(
@@ -87,7 +83,7 @@ public class FilmService {
     }
 
     public List<Film> showDirectorsFilmsAndSort(int directorId, String query) {
-        return filmStorage.findFilmsByDirectorAndSort(directorId, query);
+        return directorStorage.findFilmsByDirectorAndSort(directorId, query);
     }
 
     public List<Film> showCommonLikedFilms(int userId, int friendId) {
@@ -100,7 +96,7 @@ public class FilmService {
 
     public List<Film> showMostLikedFilmsFilter(Integer limit, Integer genreId, Integer year) {
         return filmStorage.showMostLikedFilmsFilter(limit, genreId, year);
-}
+    }
 
     public List<Film> searchFilms(String query, List<SearchingParts> by) {
         if (by.contains(SearchingParts.TITLE) && by.contains(SearchingParts.DIRECTOR)) {
