+++ src/main/java/ru/yandex/practicum/filmorate/storage/director/DirectorStorage.java
@@ -1,10 +1,8 @@
 package ru.yandex.practicum.filmorate.storage.director;
 
 import ru.yandex.practicum.filmorate.model.Director;
+import ru.yandex.practicum.filmorate.model.Film;
 
-import java.sql.Array;
-import java.sql.SQLException;
-import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 
@@ -20,4 +18,5 @@ public interface DirectorStorage {
 
     void removeDirector(long directorId);
 
+    List<Film> findFilmsByDirectorAndSort(int directorId, String query);
 }
