+++ src/main/java/ru/yandex/practicum/filmorate/storage/friend/FriendDbStorage.java
@@ -10,8 +10,6 @@ import org.springframework.stereotype.Repository;
 import ru.yandex.practicum.filmorate.mapper.Mapper;
 import ru.yandex.practicum.filmorate.model.User;
 
-import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
