+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserStorage.java
@@ -2,7 +2,6 @@ package ru.yandex.practicum.filmorate.storage.user;
 
 import ru.yandex.practicum.filmorate.model.User;
 
-import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
